% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_current_weather}
\alias{get_current_weather}
\title{title get_current_weather}
\usage{
get_current_weather(lat, lon, api_key)
}
\arguments{
\item{lat}{A number.}

\item{lon}{Another number.}

\item{api_key}{Another number.}
}
\value{
current_weather.
}
\description{
title get_current_weather
}
