% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_forecast_weather}
\alias{get_forecast_weather}
\title{title get_forecast_weather}
\usage{
get_forecast_weather(lat, lon, api_key)
}
\arguments{
\item{lat}{A number.}

\item{lon}{Another number.}

\item{api_key}{Another number.}
}
\value{
forecast_weather.
}
\description{
title get_forecast_weather
}
