% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_current_weather}
\alias{process_current_weather}
\title{process_current_weather}
\usage{
process_current_weather(current_weather)
}
\arguments{
\item{current_weather}{A number.}
}
\value{
process_current_weather.
}
\description{
process_current_weather
}
