% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{process_forecast_weather}
\alias{process_forecast_weather}
\title{process_forecast_weather}
\usage{
process_forecast_weather(forecast_weather)
}
\arguments{
\item{forecast_weather}{A number.}
}
\value{
process_forecast_weather.
}
\description{
process_forecast_weather
}
