% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{visualize_current_weather}
\alias{visualize_current_weather}
\title{visualize_current_weather}
\usage{
visualize_current_weather(current_weather_processed)
}
\arguments{
\item{current_weather_processed}{A number.}
}
\value{
visualize_current_weather.
}
\description{
visualize_current_weather
}
